Added new migration files, updated schema references, and refined configurations
