+++ README.md
@@ -1,5 +1,7 @@
 # sv
 
+VISIT: [crib-peach.vercel.app](https://crib-peach.vercel.app/)
+
 Crib is a house management app built by hackers, for hackers. Designed for shared hacker houses, Crib smooths out the chaos of eight people in a two-bedroom apartment. From voting on issues and sharing one-time secrets to sending affirmations and tracking chores — all with Discord integration — Crib brings it all together in one sleek, collaborative dashboard.
 
 HackClub Neigherbpurhood app idea: Dormroom big brother helper. log complains with an upvote downvote system, a built in app to determine what time people are free, a link to share sensative info (1 time instant link), an affirmation system. An internal member vote system to determine leader (chess elo system based on voting)
