+++ drizzle.config.ts
@@ -7,5 +7,6 @@ export default defineConfig({
 	dbCredentials: { url: process.env.DATABASE_URL },
 	verbose: true,
 	strict: true,
-	dialect: 'postgresql'
+	dialect: 'postgresql',
+	out: './drizzle'
 });
