+++ drizzle/0000_robust_molecule_man.sql
@@ -1,0 +1,25 @@
+CREATE TABLE "house_users" (
+	"id" uuid PRIMARY KEY DEFAULT gen_random_uuid() NOT NULL,
+	"created_at" timestamp DEFAULT now(),
+	"user_id" uuid NOT NULL,
+	"house_id" uuid NOT NULL
+);
+--> statement-breakpoint
+CREATE TABLE "houses" (
+	"id" uuid PRIMARY KEY DEFAULT gen_random_uuid() NOT NULL,
+	"created_at" timestamp DEFAULT now(),
+	"name" text NOT NULL
+);
+--> statement-breakpoint
+CREATE TABLE "users" (
+	"id" uuid PRIMARY KEY DEFAULT gen_random_uuid() NOT NULL,
+	"created_at" timestamp DEFAULT now(),
+	"username" text NOT NULL,
+	"email" text NOT NULL,
+	"age" integer,
+	"name" text,
+	"discord_id" text
+);
+--> statement-breakpoint
+ALTER TABLE "house_users" ADD CONSTRAINT "house_users_user_id_users_id_fk" FOREIGN KEY ("user_id") REFERENCES "public"."users"("id") ON DELETE no action ON UPDATE no action;--> statement-breakpoint
+ALTER TABLE "house_users" ADD CONSTRAINT "house_users_house_id_houses_id_fk" FOREIGN KEY ("house_id") REFERENCES "public"."houses"("id") ON DELETE no action ON UPDATE no action;
